% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.matrix.R
\name{cal.matrix}
\alias{cal.matrix}
\title{cal.matrix}
\usage{
cal.matrix(clusters, tfs)
}
\arguments{
\item{clusters}{is a list object with the clusters (maded by peaks.clustering)}

\item{tfs}{is vector of transcription factors that are contained in clusters}
}
\value{
matrix
}
\description{
Calculation of matrix for co-localized ChIP-seq peaks in clusters
}
