% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generation.random.peaks.R
\name{generation.random.peaks}
\alias{generation.random.peaks}
\title{generation.random.peaks}
\usage{
generation.random.peaks(number, ref, width = 300)
}
\arguments{
\item{number}{is amount of peaks te be generated}

\item{ref}{is a TxDb object (annotation)}

\item{width}{is a width of generated peaks}
}
\value{
GrangesObject
}
\description{
Generation random peaks with certain width
}
