% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.annotations.R
\name{peaks.annotaions}
\alias{peaks.annotaions}
\title{peaks.annotaions}
\usage{
peaks.annotaions(peaks, ref, is.Ratio = T)
}
\arguments{
\item{peaks}{is a GenomeRanges object with matadata colum that is named 'TF' (TF - Transcroption factor) and contain name for each peaks}

\item{ref}{is TxDb object}

\item{is.Ratio}{if False returns a number of peaks in diff regions, if False returns proportions of peaks in diif regions}
}
\value{
data.frame
}
\description{
Counting of peaks in genome
}
