% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.clustering.R
\name{peaks.clustering}
\alias{peaks.clustering}
\title{peaks.clustering}
\usage{
peaks.clustering(peaks, cut.off = 20, min.size = 1)
}
\arguments{
\item{peaks}{is a GenomeRanges object with matadata colum that is named 'TF' (TF - Transcroption factor) and contain name for each peaks}

\item{cut.off}{is the maximal distance between two centres of peaks that belongs to the same cluster}

\item{min.size}{is a minimal size of cluster}
}
\value{
GRangesList object where each element contains cluster of peaks
}
\description{
Find clusters based on distance between peaks
}
