% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.near.TSS.R
\name{peaks.near.TSS}
\alias{peaks.near.TSS}
\title{peaks.near.TSS}
\usage{
peaks.near.TSS(peaks, sites, range = c(-2000, 2000), ignor.strand = F,
  vertical.facet = F, horizonatl.facet = F, warp.facet = T,
  x.label = "Coordinates", y.label = "Density", legend.title = "TF",
  wide.of.line = 0.7, axis.text.size = 8, axis.title.size = 12,
  legend.title.size = 12, legend.text.size = 10, vertical.line = 0)
}
\arguments{
\item{peaks}{is a Granges object that contain TF (names of transcription factors) and Condition (colum with names of cell,tissue and so on) columns in metadata}

\item{sites}{is a number of sites (GRanges object) with the equal width, eg. promoter regions}

\item{range}{is a range with min and max values on x axis (abs sum is equale to width of sites)}

\item{ignor.strand}{is a boolean and used when strand of sites is '*'}

\item{vertical.facet}{TRUE for making vertical facet (useful when you have a lot of different TFs or Conditions)}

\item{warp.facet}{TRUE for making warp facet (useful if there are a lot of TFs)}

\item{x.label}{is a name of x axis}

\item{y.label}{is a name of y axis}

\item{legend.title}{is a title of legend}

\item{wide.of.line}{is width of curve on plot}

\item{axis.text.size}{is size of text on axis}

\item{axis.title.size}{is size of title text}

\item{legend.title.size}{is size of title text on legend}

\item{legend.text.size}{is size of text on legend}

\item{vertical.line}{is a coordinate of vertical line on plot}

\item{horizontal.facet}{TRUE for making vertical facet (useful when you have TF and Condition colums)}
}
\value{
ggplot object
}
\description{
Plot the density profile of peaks in certain sites
}
\details{
Plot the density profile of peaks in certain sites
}
