% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomizing.exp.peaks.R
\name{randomizing.exp.peaks}
\alias{randomizing.exp.peaks}
\title{randomizing.exp.peaks}
\usage{
randomizing.exp.peaks(peaks, ref)
}
\arguments{
\item{peaks}{is GRangesObject with experimental ChIP-seq peaks}

\item{ref}{is a TxDb object (annotation)}
}
\value{
GrangesObject
}
\description{
Random redistribution of ChIP-seq peaks
}
