% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.peaks.R
\name{read.peaks}
\alias{read.peaks}
\title{Peaks reading}
\usage{
read.peaks(files)
}
\arguments{
\item{peaks}{is vector of filenames (shortnames, for reading files you should be in the directory where files are located) te be readed}
}
\value{
GRangesList object where each element contains ChIP-seq peaks for certain transcription factors
}
\description{
Read .bed files in GRangesList
}
