% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/size.of.clusters.R
\name{size.of.clusters}
\alias{size.of.clusters}
\title{Calculation sizes of clusters}
\usage{
size.of.clusters(peaks, cut.off = 25)
}
\arguments{
\item{peaks}{is a GenomeRanges object}

\item{cut.off}{is the maximal distance between centres of peaks that belongs to the same cluster}
}
\value{
vector of cluster sizes
}
\description{
Calculation sizes of clusters (faster than clustering function, becouse return only vector of sizes)
}
